+++ Sources/SwiftUIKit/Blur/View+Blur.swift
@@ -21,4 +21,3 @@ public extension View {
         overlay(BlurView(style: style))
     }
 }
-