+++ Release Notes.md
@@ -3,9 +3,17 @@
 
 ## 0.2.0
 
-This release adds a `CornerRadiusStyle` style and a corresponding `cornerRadius(...)` view modifier.
+This release adds new classes and views:
 
-It also adds a `blur(...)` view modifier that simplifies using this library.
+* `CornerRadiusStyle` style can be used to specify corner radii.
+* The  `cornerRadius(_ style: CornerRadiusStyle)` view modifier can be used to apply corner radius styles.
+* `ShadowStyle` style can be used to specify drop shadow styles.
+* The  `shadow(_ style: ShadowStyle)` view modifier can be used to apply shadow styles.
+* The `blur(...)` view modifier simplifies using the system blur styles that are specified in this library.
+
+This release also adds new extensions:
+
+* The `border(content:,width:,cornerRadius)` `View` extension can be used to apply borders with a certain content (e.g. Color.red), width and corner radius.
 
 
 ## 0.1.0